+++ src/balancing_platform/__init__.py
@@ -4,6 +4,7 @@
 
 
 from pkg_resources import get_distribution, DistributionNotFound
+import scipy
 import cv2
 import numpy
 import pymodbus
@@ -13,7 +14,7 @@ import matplotlib
 try:
     # Change here if project is renamed and does not equal the package name
     dist_name = __name__
-    __all__ = [cv2, numpy, pymodbus, vpython, matplotlib]
+    __all__ = [scipy, cv2, numpy, pymodbus, vpython, matplotlib]
     __version__ = get_distribution(dist_name).version
 except DistributionNotFound:
-    __version__ = '0.6'
+    __version__ = '0.7'