+++ src/balancing_platform/util/graphs.py
@@ -1,0 +1,10 @@
+#!/usr/bin/env python
+# -*- coding: utf-8 -*-
+"""
+
+Code by: Magnus Øye, Dated: 12.11-2018
+Contact: magnus.oye@gmail.com
+Website: https://github.com/magnusoy/Balancing-Platform
+"""
+
+# Importing packages